Rework PCP writer stuff to be less File centric, to allow in memory ByteBuffer creation for unit testing purposes, plus also allow a Pcp Writer to restart cleanly if new Metrics arrive.
Add to SelfStartingMonitoringViewTest an integration test that tests the way the PCPMonitorBridge actually resets and restarts when new metrics are added later.  It is probably better for this to be included in an overarching integration-test module; in fact the parfait-spring dependency onto parfait-pcp is left at test only scope for this reason.

Astonishingly, all the unit tests are passing now.  Bit spooky if you ask me.