Allow optional managed bean definitions, add metaspace metrics
The permanent generation metrics for the various garbage collection
implementations have had problems related to whether or not metrics
exist within a JVM.  Original parfait handles the case where permgen
metrics is one of a set of named beans, but not the case from Java 8
where permgen is done away with altogether (replaced with metaspace,
and the alternate G1 algorithms).

This was a problem originally, as we attempted to define spring beans
that didn't exist.  We now introduce a mechanism for optional metrics
and use this to (re)introduce permgen and metaspace metrics.