Shuffle forward to using current spring version (4.2.5)
We need a newer Spring for incoming parfait-agent Spring testing.
Updates needed to one of the CXF tests, where the new version was
seeing a FileNotFound exception - traced back to use of port 9080
for the test-created http server, which generates an exception if
one reads from the output stream rather than error stream.