Begin using the spring-test framework in ParfaitAgent
This required a bump to the Junit version beyond 4.9 - went with
current latest which is 4.12 for now.

Agent test added which verifies that all the spring bean loading
activity which -javaagent mode would invoke happens without any
errors, and the DynamicMonitoringView is successfully created.

The test uses a custom properties file for injecting the needed
parfait-agent-specific properties, which meant augmenting the bean
configuration with an additional file to load in custom properties
- hence the @ContextConfiguration test class annotation lists both
of those files.