MMVv2: Introduce MmvVersion enum with factory methods
The MmvVersion enum is intended to be used publicly when constructing an
instance of PcpMmvVersion. It maintains backwards-capability by defaulting
to version 1. We might want another constructor when using a File and an
IdentifierSourceSet. This can be added in easily if needed.

It also operates as a factory to create the correct version of the validator
and store classes. Its a bit of extra behaviour but it constrains all the
version-specificness decision making to this enum only.