Replace monitorable AgentMonitoringView fields with parameters
Fix createMonitorable use of mBeanName and attributeName fields:
because of the way metrics were now being registered (now not by
Spring), it was changing that field each time instead of creating
completely new objects.

This resulted in a series of PollingMonitoredValues all accessing
the same field.  It was working in Spring because the definition
was abstract, so each bean got created with its own instance of
these fields.

Fix this by replacing those fields with parameters and the values
correctly update once more.