Transition parfait agent and core to using slf4j exclusively
We had previously a handful of places using Log4J explicitly;
these are now transitioned to use the Slf4J wrappers only and
in the case of parfait-agent defaulting to slf4j-simple (when
no other logger is available / configured).

This change removes last dependencies on Log4J in Parfait - a
developer can still choose to use Log4J with Parfait of course
but that is now an explicit, opt-in choice.

Resolves Red Hat BZ #2031674