fix PcpMmvWriter strings leak
`PcpMmvWriter` leaks strings through its `PcpStringStore` when the
writer is reset. The fix is to clear the store in the reset method.
However, the `metricInfoStore` and `instanceDomainStore` rely on
putting values in the string store when MMV2 is used. As such, these
stores must also be cleared so that the strings can be recreated.