Fix race updating metrics with writer is stopped
This fixes another race condition that causes metrics to be lost when
they are updated while the writer is stopped and after the monitorable
has been re-added to the writer. In this case, the update needs to set
the initial value on `PcpValueInfo` so that the correct value is
written when the writer is started.