Add caching to improve speed adding monitorables
Adds a cache to the `MetricNameMapper` and `MetricNameValidator`. This
improves performance of mapping metric names because a regex no longer
has to be executed over and over again. The validator cache helps
because the metric name no longer has to be encoded over and over
again.

Neither of these caches expire entries, but this is not an issue
because it's currently not possible to remove monitorables from
Parfait.